backend/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
pymupdf
anthropic
openai
pydantic
psycopg2-binary
python-dotenv
